Move requirements from route blinding section
The route blinding section should only contain requirements to help encrypt
and decrypt blinded paths. Any requirement that is specific to payment
should be in the payment onion payload, as suggested by @rustyrussell.

Future requirements for how to use blinded paths in onion messages will be
included in the onion messages requirements, instead of the route blinding
ones.